requirements.txt
streamlit>=1.27.2
joblib>=1.3.2
pandas>=2.0.3
scikit-learn>=1.2.2
requests>=2.31.0
hugchat>=0.3.1
Pillow>=10.0.0
numpy>=1.24.0
reportlab>=4.0.0
torch>=2.2.0
torchvision>=0.17.0
matplotlib>=3.7.0
